#75 - refactor battlestage ui
