+++ Assets/Scripts/BattleStage/BattleStageUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 55001c68b12dc44febc9abcc1569f3d7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
