+++ Assets/Scripts/DOWGameManager.cs
@@ -63,7 +63,7 @@ namespace DOW
 
             #region Manager 세팅
             instance.AddManager(typeof(TimeManager), TimeManager.Instance, true);
-            Instance.AddManager(typeof(SoundManager), SoundManager.Instance, false);
+            instance.AddManager(typeof(SoundManager), SoundManager.Instance, false);
             instance.AddManager(typeof(TableManager), TableManager.Instance, false);
             instance.AddManager(typeof(PopupManager), PopupManager.Instance, false);
             instance.AddManager(typeof(UIManager), UIManager.Instance, false);
