+++ Assets/Scripts/StartMenu/StartMenuUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 970e621845c694401a8ff56e725a2956
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
